+++ sample/src/app.service.ts
@@ -1,6 +1,6 @@
-import { GraphileWorkerListener, OnWorkerEvent } from '@app/graphile-worker';
 import { Injectable, Logger } from '@nestjs/common';
 import { WorkerEventMap } from 'graphile-worker';
+import { GraphileWorkerListener, OnWorkerEvent } from '../../src/index';
 
 @Injectable()
 @GraphileWorkerListener()
@@ -10,5 +10,10 @@ export class AppService {
   @OnWorkerEvent('job:success')
   onJobSuccess({ job }: WorkerEventMap['job:success']) {
     this.logger.debug(`job #${job.id} finished`);
+  }
+
+  @OnWorkerEvent('job:error')
+  onJobError({ job, error }: WorkerEventMap['job:error']) {
+    this.logger.error(`job #${job.id} fail ${JSON.stringify(error)}`);
   }
 }