+++ sample/src/hello.task.ts
@@ -1,10 +1,6 @@
-import { Task, TaskHandler } from '@app/graphile-worker';
 import { Injectable, Logger } from '@nestjs/common';
 import { Helpers } from 'graphile-worker';
-
-export function helloTask(payload: any, helpers: Helpers) {
-  console.log(`hello task %o`, payload);
-}
+import { Task, TaskHandler } from '../../src/index';
 
 @Injectable()
 @Task('hello')
@@ -13,6 +9,6 @@ export class HelloTask {
 
   @TaskHandler()
   handler(payload: any, _helpers: Helpers) {
-    this.logger.log(`handle ${payload}`);
+    this.logger.log(`handle ${JSON.stringify(payload)}`);
   }
 }