+++ src/index.ts
@@ -1,3 +1,4 @@
-export * from './decorators/worker-hooks.decorators';
+export * from './decorators/task.decorators';
+export * from './decorators/worker.decorators';
 export * from './graphile-worker.module';
 export * from './services/worker.service';