+++ src/services/metadata-accessor.service.ts
@@ -1,9 +1,14 @@
 import { Injectable, Type } from '@nestjs/common';
 import { Reflector } from '@nestjs/core';
 import {
-  GRAPHILE_WORKER_LISTENER,
-  GRAPHILE_WORKER_ON_WORKER_EVENT,
-} from '../decorators/worker-hooks.decorators';
+  TASK_HANDLER_METADATA,
+  TASK_METADATA,
+} from '../decorators/task.decorators';
+import {
+  WORKER_LISTENER_METADATA,
+  WORKER_ON_EVENT_METADATA,
+} from '../decorators/worker.decorators';
+import { WorkerEventName } from '../interfaces/worker.interfaces';
 
 /**
  * Heavily inspired from [`BullMetadataAccessor`](https://github.com/nestjs/bull/blob/c230eab1dc26fb743a3428e61043167866b1e377/lib/bull-metadata.accessor.ts)
@@ -12,23 +17,41 @@ import {
 export class MetadataAccessorService {
   constructor(private readonly reflector: Reflector) {}
 
-  isListener(target: Type<any> | Function): boolean {
-    if (!target) {
-      return false;
-    }
+  // WORKER DECORATOR
 
-    return !!this.reflector.get(GRAPHILE_WORKER_LISTENER, target);
+  isListener(target: Type<any> | Function): boolean {
+    return this.hasMetadata(WORKER_LISTENER_METADATA, target);
   }
 
   isWorkerEvent(target: Type<any> | Function): boolean {
+    return this.hasMetadata(WORKER_ON_EVENT_METADATA, target);
+  }
+
+  getListenerMetadata(
+    target: Type<any> | Function,
+  ): WorkerEventName | undefined {
+    return this.reflector.get(WORKER_ON_EVENT_METADATA, target);
+  }
+
+  // TASK DECORATOR
+
+  isTask(target: Type<any> | Function): boolean {
+    return this.hasMetadata(TASK_METADATA, target);
+  }
+
+  isTaskHandler(target: Type<any> | Function): boolean {
+    return this.hasMetadata(TASK_HANDLER_METADATA, target);
+  }
+
+  getTaskMetadata(target: Type<any> | Function): string | undefined {
+    return this.reflector.get(TASK_METADATA, target);
+  }
+
+  private hasMetadata(key: Symbol, target: Type<any> | Function): boolean {
     if (!target) {
       return false;
     }
 
-    return !!this.reflector.get(GRAPHILE_WORKER_ON_WORKER_EVENT, target);
-  }
-
-  getListenerMetadata(target: Type<any> | Function): any | undefined {
-    return this.reflector.get(GRAPHILE_WORKER_ON_WORKER_EVENT, target);
+    return !!this.reflector.get(key, target);
   }
 }