+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Current
 
+- **feat(task)**: provide a `@Task(name)` decorator to define your injectable tasks
 - **fix(hooks)**: Set all worker hooks (was only `job:success`)
 - **refactor(configuration)**: Remove `ConfigurationService` class and use a plain object `RunnerOptions`
 - **refactor()**: Rename `GraphileWorkerService` to `WorkerService`